LayoutTests/inspector/timeline/timeline-mark-timeline.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/timeline-test.js"></script>
<script>

function performActions()
{
    console.markTimeline("MARK TIMELINE");
}

function test()
{
    InspectorTest.performActionsAndPrint("performActions()", "TimeStamp");
}

</script>
</head>

<body onload="runTest()">
<p>
Tests the Timeline API mark feature
</p>

</body>
</html>
